:construction_worker: add linter

+++ src/Domain/Budget/Model/BudgetStatusEnum.php
@@ -4,15 +4,15 @@ namespace App\Domain\Budget\Model;
 
 enum BudgetStatusEnum: string
 {
-    case STATUS_HAS_MORE_SAVE = 'has-more-save';
+    case STATUS_HAS_MORE_SAVE  = 'has-more-save';
     case STATUS_IS_IN_PROGRESS = 'is-in-progress';
-    case STATUS_IS_SOON_FULL = 'is-soon-full';
-    case STATUS_IS_FULL = 'is-full';
-    case STATUS_IS_OVER = 'is-over';
+    case STATUS_IS_SOON_FULL   = 'is-soon-full';
+    case STATUS_IS_FULL        = 'is-full';
+    case STATUS_IS_OVER        = 'is-over';
 
-    public static function statusByProgress(int $relativeProgress): self
+    public static function statusByProgress(float $relativeProgress): self
     {
-        if (0 === $relativeProgress) {
+        if (0.0 === $relativeProgress) {
             return self::STATUS_HAS_MORE_SAVE;
         }
 
@@ -20,7 +20,7 @@ enum BudgetStatusEnum: string
             return self::STATUS_IS_SOON_FULL;
         }
 
-        if (100 === $relativeProgress) {
+        if (100.0 === $relativeProgress) {
             return self::STATUS_IS_FULL;
         }
